Added skeleton of interpolation using ADBs as input
